Move Create Tenant to Thunk
