+++ portal-ui/src/screens/Console/Tenants/types.ts
@@ -384,5 +384,3 @@ export interface ITenantIdentityProviderResponse {
     user_dn_search_filter: string;
   };
 }
-
-export type FieldsToHandle = INameTenantFields;
